Overhauled documentation for readability and added support section.
I added clearer examples differentiating the callback and Promise forms.  Hopefully, this will make the doc read a little better.  In addition, showed both use patterns in the example to exemplify the ease of migration when using this library.

Lastly, I've added a small Support section.  This describes what browser and version this library is intended to support with relevant specifications.  Basically, this library is intended to support any Chrome version that supports native Promises.  There are hundreds of Promise libraries and shims at this point but any of them that respect the resolver/rejector pattern can be used if one really needs to support a Chrome version pre 33.  Keep in mind however, that many neat and useful Chrome APIs are unavailable before v33 so it may not be worthwhile to use fancy promises when upgrading the environment may be a much more appropriately low-hanging fruit.